// src/frontend/src/flatbuffers/base/parameter-type.ts
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

export enum ParameterType {
  UNKNOWN = 0,
  STRING = 1,
  FLOAT = 2,
  INTEGER = 3,
  BOOLEAN = 4,
  ARRAY = 6
}